Moving things into a module
Adds mini tests
